% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{build_model}
\alias{build_model}
\title{Build and Train Poisson Model, Save to Disk}
\usage{
build_model()
}
\value{
The path to the saved model file.
}
\description{
This function orchestrates the entire process of model building:
- Initializes the temporary workspace,
- Extracts and transforms the data,
- Splits the data into training, validation, and test sets,
- Trains the model using the training pipeline,
- Saves the trained model to disk.
}
\examples{
\dontrun{
Example usage:
model_path <- build_model()
}
}
