% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{elt_to_temp_dir}
\alias{elt_to_temp_dir}
\title{Export JSON to Temporary Directory as Parquet}
\usage{
elt_to_temp_dir(file_name)
}
\arguments{
\item{file_name}{A string representing the base name of the file to process. The JSON file is expected to be located in the S3 bucket with the given name.}
}
\value{
A string containing the file path of the generated Parquet file in the temporary directory.
}
\description{
This function connects to DuckDB, installs and loads necessary extensions,
and converts a remote JSON file to a Parquet file saved in the temporary directory.
}
\examples{
\dontrun{
elt_to_temp_dir("match_results")
}
}
