% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{extract_data}
\alias{extract_data}
\title{Extract Data from S3 and Return Datasets}
\usage{
extract_data()
}
\value{
A named list containing:
\describe{
  \item{match_results}{A tibble with match results.}
  \item{innings_results}{A tibble with innings results.}
}
}
\description{
This function extracts datasets from remote JSON files, processes them into
Parquet format, reads the data, and removes temporary files. It returns
the data as a list of tibbles.
}
\examples{
\dontrun{
data <- extract_data()
}
}
