% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_temp_workspace}
\alias{initialize_temp_workspace}
\title{Initialize Temporary Workspace}
\usage{
initialize_temp_workspace()
}
\value{
A list with two elements: `staging` and `model`, each being the path to
  the respective directory.
}
\description{
Creates a temporary directory structure with "staging" and "model" subdirectories
inside the base temporary directory. The function returns a list containing the paths
to these directories for further use in model storage and staging purposes.
}
\examples{
\dontrun{
temp_dirs <- initialize_temp_workspace()
temp_dirs$staging  # Path to the staging directory
temp_dirs$model    # Path to the model directory
}
}
