% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{split_data}
\alias{split_data}
\title{Split Data into Training, Validation, and Test Sets}
\usage{
split_data(df, prediction = FALSE)
}
\arguments{
\item{df}{A data frame to be split.}

\item{prediction}{A logical flag indicating whether to return the modified data frame without splitting. Default is FALSE.}
}
\value{
A list containing three elements:
  - `train_data`: Data frame for training.
  - `val_data`: Data frame for validation.
  - `test_data`: Data frame for testing.
  If `prediction` is TRUE, the modified data frame is returned instead.
}
\description{
This function splits a given data frame into three subsets: training (60%), validation (20%), and test (20%).
It removes the 'team' column from the data frame before performing the split. If the `prediction` argument is set to TRUE,
it returns the modified data frame without performing the split.
}
\examples{
\dontrun{
# Example usage:
result <- split_data(my_data)
training_set <- result$train_data
validation_set <- result$val_data
testing_set <- result$test_data
}
}
