% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transform_data}
\alias{transform_data}
\title{Transform Match and Innings Data}
\usage{
transform_data(datasets)
}
\arguments{
\item{datasets}{A list containing:
\itemize{
  \item `match_results`: A tibble of match results.
  \item `innings_results`: A tibble of innings results.
}}
}
\value{
A tibble containing the aggregated innings data with the following columns:
  \itemize{
    \item `team`: Team name.
    \item `innings`: Inning number.
    \item `overs_remaining`: Overs remaining for the team.
    \item `wickets_remaining`: Wickets remaining for the team.
    \item `runs`: Runs scored by the team.
  }
}
\description{
Processes the input datasets containing match and innings data to create a transformed dataset
by joining and aggregating the information.
}
